scripts refactor
